feat: Forms UI
Here is a try to make the forms UI.
I dont know if this is the right way to do thing or not since im not that experienced in React, but it still a try, u can judge, or maybe take whats good to use.
here i did the Help request and offer forms, their text inputs, text area inputs, file inputs, type inputs, no phone inputs because you will most likely install some package and i will let that to you, Map input still needs work, i set it with a dragable marker for now, but i would imagine it should place marker and get info on click.