feat: infinite scrolling with pull to refresh
A first implementation of the infinite scrolling feature from scratch with pull to refresh package https://www.npmjs.com/package/react-pull-to-refresh, for a dummy list of 1000 elemets, it just needs the loading spinner or a loading component.

Note that i tried the react-infinite-scroll-component package, but it seems to not work for some reason.

Other changes are style changes you may or may not disregarded white resolving conflicts, until u resolve em again :)