feat: map view switch button & custom zoomIn and zoomOut buttons #43
Added a switch to toogle between the view in the map, and since the zoom buttons cant be centered in the leaflet map, i made custom buttons and positioned them the way i want.
Design is reviewed by Maadlou.
Fixed a small bug in the nav bar, see the pictures attached to see the difference.